Add ability to search on country and coordinates
Split the search term with a comma and the first part is used to search
a city and the second part is used to filter a country code. It is
also possible to search on latitude and longitude. Also use a comma
as a separator for this. The first part is used for lat, the second
part for lon.